+++ app/build.gradle.kts
@@ -2,6 +2,10 @@ plugins {
     alias(libs.plugins.android.application)
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.kotlin.compose)
+    alias(libs.plugins.dagger.hilt.plugin)
+    alias(libs.plugins.kotlin.kapt)
+    alias(libs.plugins.serialization)
+    //alias(libs.plugins.google.services)
 }
 
 android {
@@ -51,6 +55,23 @@ dependencies {
     implementation(libs.androidx.activity.compose)
     // Coil image loading library for Compose to load and display images efficiently
     implementation(libs.io.coil.kt.compose)
+
+    //Dagger Hilt
+    implementation(libs.dagger.hilt)
+    kapt(libs.dagger.hilt.compiler)
+    implementation(libs.androidx.hilt.navigation.compose)
+
+    //Navigation compose
+    implementation(libs.androidx.navigation.compose)
+
+
+    //firebase
+
+//    implementation(platform(libs.google.firebase.bom))
+//    implementation(libs.google.firebase.analytics.ktx)
+//    implementation(libs.google.firebase.auth.ktx)
+//    implementation(libs.google.firebase.firestore.ktx)
+
     implementation(platform(libs.androidx.compose.bom))
     implementation(libs.androidx.ui)
     implementation(libs.androidx.ui.graphics)
