+++ app/src/main/java/com/example/ahyaha/AhyahaApp.kt
@@ -1,0 +1,8 @@
+package com.example.ahyaha
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+
+@HiltAndroidApp
+class AhyahaApp : Application()
