+++ app/src/main/java/com/example/ahyaha/MainActivity.kt
@@ -8,12 +8,15 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
-import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.ahyaha.ui.theme.AhyahaTheme
-import com.example.ahyaha.view.MainScreen
-import com.example.ahyaha.viewmodel.BloodTypeViewModel
-import com.example.ahyaha.viewmodel.DonorViewModel
+import com.example.ahyaha.presentation.view.MainScreen
+import com.example.ahyaha.presentation.viewmodel.BloodTypeViewModel
+import com.example.ahyaha.presentation.viewmodel.DonorViewModel
+import dagger.hilt.android.AndroidEntryPoint
 
+
+@AndroidEntryPoint
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -22,8 +25,8 @@ class MainActivity : ComponentActivity() {
             AhyahaTheme {
                 Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
 
-                    val donorViewModel = viewModel<DonorViewModel>()
-                    val bloodTypeViewModel = viewModel<BloodTypeViewModel>()
+                    val donorViewModel = hiltViewModel<DonorViewModel>()
+                    val bloodTypeViewModel = hiltViewModel<BloodTypeViewModel>()
 
 
                     MainScreen(
