+++ app/src/main/java/com/example/ahyaha/data/model/Donor.kt
@@ -1,4 +1,4 @@
-package com.example.ahyaha.model
+package com.example.ahyaha.data.model
 
 import java.util.Date
