+++ app/src/main/java/com/example/ahyaha/domain/usecase/GetDonorsUseCase.kt
@@ -1,0 +1,8 @@
+package com.example.ahyaha.domain.usecase
+
+import com.example.ahyaha.data.repository.DonorRepository
+import javax.inject.Inject
+
+class GetDonorsUseCase @Inject constructor(private val donorRepository: DonorRepository) {
+    operator fun invoke() = donorRepository.getAllDonors()
+}
