--- app/src/main/java/com/example/ahyaha/view/MainScreen.kt
@@ -1,76 +1,0 @@
-package com.example.ahyaha.view
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Modifier
-import com.example.ahyaha.viewmodel.BloodTypeViewModel
-import com.example.ahyaha.viewmodel.DonorViewModel
-import android.annotation.SuppressLint
-import androidx.compose.foundation.*
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.padding
-import androidx.compose.runtime.Composable
-
-@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
-@Composable
-fun MainScreen(
-    donorViewModel: DonorViewModel,
-    bloodTypeViewModel: BloodTypeViewModel,
-    modifier: Modifier = Modifier
-) {
-    val donorState by donorViewModel.uiState.collectAsState()
-    var selectedTab by remember { mutableStateOf(0) }
-    var searchText by remember { mutableStateOf("") }
-
-    val filteredDonors = donorState.donors.filter {
-        it.name.contains(searchText, ignoreCase = true)
-    }
-
-    Scaffold(
-        bottomBar = { BottomNavigationBar { selectedTab = it } }
-    ) { innerPadding ->
-        Box(modifier = Modifier.padding(innerPadding)) {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .verticalScroll(rememberScrollState())
-            ) {
-                TopBar(searchText) { searchText = it }
-                BloodTypesSection()
-                ImageSection()
-                RegularDonorsSection(donors = filteredDonors)
-                Events()
-                ActivitySection()
-                RecentPostsSection()
-            }
-        }
-    }
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
