--- app/src/main/java/com/example/ahyaha/viewmodel/DonorViewModel.kt
@@ -1,40 +1,0 @@
-package com.example.ahyaha.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.example.ahyaha.model.Donor
-import com.example.ahyaha.repository.DonorRepository
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.launch
-
-data class UIState(
-    val isLoading: Boolean = false,
-    val donors: List<Donor> = emptyList(),
-    val error: String? = null
-)
-
-class DonorViewModel : ViewModel() {
-    
-    private val _uiState = MutableStateFlow(UIState())
-    val uiState: StateFlow<UIState> = _uiState.asStateFlow()
-
-    init {
-        getDonors()
-    }
-
-    fun getDonors() {
-        viewModelScope.launch {
-            _uiState.value = _uiState.value.copy(isLoading = true)
-            try {
-                delay(2000)
-                val donors = DonorRepository.getAllDonors()
-                _uiState.value = _uiState.value.copy(donors = donors, isLoading = false)
-            } catch (e: Exception) {
-                _uiState.value = _uiState.value.copy(error = e.message, isLoading = false)
-            }
-        }
-    }
-}
