+++ build.gradle.kts
@@ -3,4 +3,7 @@ plugins {
     alias(libs.plugins.android.application) apply false
     alias(libs.plugins.kotlin.android) apply false
     alias(libs.plugins.kotlin.compose) apply false
+    alias(libs.plugins.dagger.hilt.plugin) apply false
+    alias(libs.plugins.kotlin.kapt) apply false
+    alias(libs.plugins.google.services) apply false
 }
